+++ tests/functional/test_station.py
@@ -51,11 +51,12 @@ def test_tile(stations_devices_exported: list[tango.DeviceProxy]) -> None:
 
 
 @given("an SPS deployment against HW")
-def check_against_hardware(hw_context: bool) -> None:
+def check_against_hardware(hw_context: bool, station_label: str) -> None:
     """
     Skip the test if not against HW.
 
     :param hw_context: whether or not the current context is against real HW.
+    :param station_label: Station to test against.
     """
     if not hw_context:
         pytest.skip("This test requires real HW.")